refactor: ui/list 패키지 클래스와 ui/today 패키지 클래스 ui/ lesson 패키지로 병합
